Fix "checkout" used as a verb to "check out"
There is no verb "to checkout" in English. If there was, we could
write things like "the file has been checkouted" or "checkouting
branch X failed". Clearly that's not English :-) The verb is simply
"to check out", a file has been "checked out", and "checking out" a
branch might fail.

Change-Id: I69d0becac28fd144e2e16d59720975950e3b062a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>