Staging View tooltips: work around bug 330384 on GTK
On GTK at least, the artificial inset of 5 pixels used in the
tooltip's internal TooltipHideListener closes the tooltip before
the mouse ever can move over the tooltip. It appears that on GTK
the tooltip's shell gets an SWT.MouseExit event with coordinates
inside the tooltip, but not within the smaller inset area.

It's utterly unclear to me why this inset was ever programmed anyway.
Removing it and always testing against the true bounds of the tooltip
resolves the problem and makes it possible to move the mouse cursor
over the tooltip, also on GTK.

The behavior on MacOS is unchanged.

Bug: 516969
Change-Id: I009be811e0fee9483ed7a0827b5dfe15c4d4490f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>