FetchGerritChangePage: auto-supply highest patch set number
If the change field contains only a change number but no patch set
number, automatically fill in the highest patch set number obtained
from Gerrit once we have the results of the ls-remote. If the field
is the focus control and its contents have changed since we started
the content assist, the user might be typing: don't interfere and do
not auto-update the field.

Bug: 518492
Change-Id: I5fa348aed4b6881c87749bbed6789f97d3e4e1d9
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>