Improve hard-wrapping in commit messages
Preserve the current behavior as much as possible, but relax the
rules about wrapping long words. This prevents moving long URLs
(or other long "words") to new lines when doing so doesn't improve
the formatting. In particular, it'll keep URLs used in references[1]
on the line with the reference index.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=519012#c0&dummy=0&dummy=0&dummy=0

Bug: 519012
Change-Id: Ic13662d6c3eed93957d9486a88fe1c92a1157019
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>