Fix alignment of buttons in git config preference pages
The "Open" button (and the composite containing it) had more margin than
all the other widgets on the preference page. This change makes them all
perfectly aligned.

Removed a non-used 4th column in the location composite layout and
removed the default 5px marginWidth that is created for every
GridLayout.

Tested on Windows only, but this should not be Windows specific.

Change-Id: I701f0cdbdb8465c99fb336f99ff4e776b7d444a8
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>