Add descriptive message to directory browsers
Directory browsers allow showing a short message about the current
operation. This can be helpful if multiple directory browsers are opened
or after coming back to a directory browser after a task switch.

This change adds such a message to all directory browsers in egit.
Typically the message is the same as the title of the wizard using the
directory browser.

Change-Id: If680ce115895fd8ca56e0ed337c6bba6e3de5855
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>