Ignore invalid .project files when importing projects
When a ProjectRecord has no ProjectDescription, this indicates that
an error occurred while trying to read the .project file. EGit in that
case tried to be smart about it and to supply a new ProjectDescription,
but that only lead to follow-up errors.

Instead ignore such ProjectRecords completely. The platform will
already have logged an error about it (and we can't do anything about
that), but at least we can avoid even considering such projects for
import.

In case some such ProjectRecord should still get to
ProjectUtils.createExistingProject(), log an error with the offending
file path, but otherwise just don't do anything.

Bug: 521745
Change-Id: I09bdcf22390fd952fb2c66163252aa5d337dae90
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>