Provide property name for repository tree nodes
Provide a label to be used for all different kinds of repository tree
nodes, not only for the root repository node. This makes the IDE
property dialog display the "No properties" dialog with the correct node
label.

This patch does not really add new functionality, it only makes the
already implemented adapter available for more classes in the same
hierarchy.

Bug: 527068
Change-Id: I2fbde50320b9d785a219a00c5defa27952286616
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>