Make the PushWizardDialog a NonBlockingWizardDialog
No functional change yet; we're not using the non-blocking features
yet.

Change the NonBlockingWizardDialog such that the inherited run()
is available to get the normal blocking modal behavior. The only
place where we do use the non-blocking run() uses the new method
anyway, and in the multi-page PushBranchWizard, we might have
other use cases where we'd still need to use the modal run().

Change-Id: Id8edfc577367d005abab77828ae7ebb6c5081bf9
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>