Keep Gerrit Change-Id for conflicting cherry-pick
Gerrit itself does so, too, when cherry-picking in its web UI.

Keeping the Change-Id will result in a new patch set if the
cherry-picked commit has been fetched from Gerrit, and if that
Gerrit change is on the same upstream branch as the target of
the cherry-pick.

With a new Change-Id, pushing to Gerrit would create a new initial
patch set for a new change in any case.

Bug: 533336
Change-Id: I20905c86ea9a6e9137a7d6b4f8d408a95510ae83
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>