Disable commit buttons immediately on committing
Since all the file related checks can take quite some time, we need to
disable the buttons immediately, not only when triggering the commit
operation on the repository.

Most notable visible difference is that the complete staging view is
disabled when the user gets the "no files staged" error message dialog.

Change-Id: Iae825abfb9235832a3fae49e9f967d983461dec5
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>
Bug:533005