Fix wrapping label in DateFormatPreferencePage
Using SWT.FILL and SWT.TOP as _width and height_ makes no sense.
The real problem with wrapping labels in a GridLayout is that one
needs some width to wrap against. With SWT.DEFAULT, the layout lets
the label itself determine its width, which of course will always
be the unwrapped width absent any other constraints.

If one wants to have wrapping, one needs to provide a reasonable
widthHint.

Change-Id: Ie07d2fd93f110048ddcfb535d0706347915852d0
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>