Unify order of branch context menu items
The order of the menu items Merge, Rebase, Reset, Synchronize in the
repository view context menu is different, depending on whether the menu
is opened on the currently checked out branch or a not checked out
branch.

This change modifies the context menu for not checked out branches to be
more similar to the one for the checked out branch.

Bug: 538844
Change-Id: I19687bd5ca08b4eb1fcb5ba5f943bdc2113baf36
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>