Don't log an error on save in comparisons if resource out of sync
CompareEditorInput.saveChanges() also only displays a dialog if
something goes wrong with saving. We cannot just let the exception
propagate because this is also reachable via other code paths that
do not appear to check for InvocationTargetException. So we need to
handle it somehow.

Do not log the error; just present an error to the user via the
StatusManager.

Bug: 538855
Change-Id: I03f0699f4641a9ba99815aad0869c0ef1d4c2a66
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>