Use a separate job family for FileDiffContentProvider
Using the HISTORY_DIFF family was wrong; it's reserved for generating
the unified diff. That family is canceled in GitHistoryPage when new
diff are to be generated, which may inadvertently also cancel the file
diff generation.

Give the job in the asynchronous FileDiffContentProvider its own job
family. It needs a family so that UI tests can join on it.

Bug: 539099
Change-Id: I6f168cbd2544d2e08617536271a71ed6ebd027fb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>