ReflogView: update only for ref changes in current repository
The reflog view uses a global RefsChangedListener, which gets called
whenever any ref in any repository changes. Make sure it updates only
if the event is for the repository currently begîng shown.

Also handle the case when the currently shown ref has been deleted:
fall back to showing the reflog for HEAD.

Change-Id: I692818576fd5131e42e3a82919e3fd5d5cd2ca43
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>