Remove numbers from "Copy X Paths" menu item
In a previous change I converted more strings to use pluralization
instead of 2 separate entries, sometimes also adding the number as such
in the localization. For this menu item that looks bad in practice,
since that is too much information in the menu item. Therefore revert
using the number here.

Sorry for the noise.

Change-Id: I8c0440f24f4220b75d8b6ad45acd1c9001b5f400
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>