EGit should respect the SSH2 home directory from the preferences
Changes to the SSH2 home directory in the Eclipse preferences were
ignored by EGit; it always used ~/.ssh as SSH2 home directory. That's
not how it was intended.

However, note that setting the SSH2 home directory to anything but
~/.ssh in the Eclipse preferences is likely to be very confusing.
EGit will then look for the known_hosts file and the SSH config
file in that directory, and will also look there for the default
identity files. EGit will then be using a setup different from just
about all other SSH tools, and things that work in Eclipse may not
work on the command-line (or in other tools) or vice versa.

Nevertheless: given that the possibility of changing that directory
exists at all on that preference page, EGit should respect it.

Bug: 543855
Change-Id: Ib95741c209956f3883891e1578c56e6535dbc8a2
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>