[history] Handle workspace and non-workspace files alike
With compare mode on and the history view filtered to one file, a
double click on a commit opens a compare editor for that file. But
the behavior for files in the Eclipse workspace and files outside
of the Eclipse workspace were different:

* IResource (file in workspace): compare against working tree
* File (file not in Eclipse workspace): compare against HEAD

The context menu also had the same difference.

Judging from the (now removed) TODO in CompareWithWorkingTreeHandler
comparing against HEAD was just a work-around for not being able to
create a compare editor input for plain java.io.Files.

Since LocalNonWorkspaceTypedElement enables us to do exactly that,
use it and also compare non-workspace files against the working tree.

Remove the Compare with HEAD menu command from the history view
context menu completely. Remove the handler, too, and simply always
use "Compare with Working Tree" instead. "Compare with HEAD" is
available anyway by selecting the HEAD commit and another commit,
then choosing "Compare with each other".

Change-Id: Ic04168432cf64928aa73b81d3cf250f096e74b31
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>